+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/RequestVerifier.kt
@@ -13,7 +13,7 @@ import org.codice.compliance.SAMLCore_3_2_1_c
 import org.codice.compliance.SAMLCore_3_2_1_e
 import org.codice.compliance.attributeNode
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_3_2
+import org.codice.compliance.Section.CORE_3_2
 import org.codice.compliance.utils.CONSENT
 import org.codice.compliance.utils.DESTINATION
 import org.codice.compliance.utils.ID
