+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/ResponseVerifier.kt
@@ -18,7 +18,7 @@ import org.codice.compliance.attributeText
 import org.codice.compliance.children
 import org.codice.compliance.recursiveChildren
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_3_2
+import org.codice.compliance.Section.CORE_3_2
 import org.codice.compliance.utils.CONSENT
 import org.codice.compliance.utils.DESTINATION
 import org.codice.compliance.utils.ID
