+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/SignatureSyntaxAndProcessingVerifier.kt
@@ -12,7 +12,7 @@ import org.codice.compliance.SAMLCore_5_4_2_a
 import org.codice.compliance.attributeText
 import org.codice.compliance.recursiveChildren
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_5_4
+import org.codice.compliance.Section.CORE_5_4
 import org.codice.compliance.utils.ID
 import org.w3c.dom.Node
