+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/requests/CoreLogoutRequestProtocolVerifier.kt
@@ -11,7 +11,7 @@ import org.codice.compliance.SAMLCore_3_7_1_a
 import org.codice.compliance.SAMLCore_3_7_3_2_e
 import org.codice.compliance.attributeNode
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_3_7
+import org.codice.compliance.Section.CORE_3_7
 import org.codice.compliance.utils.NodeDecorator
 import org.codice.compliance.verification.core.CommonDataTypeVerifier
 import org.codice.compliance.verification.core.RequestVerifier
