+++ ctk/common/src/main/kotlin/org/codice/compliance/verification/core/responses/CoreLogoutResponseProtocolVerifier.kt
@@ -12,7 +12,7 @@ import org.codice.compliance.SAMLCore_3_7_3_2_d
 import org.codice.compliance.attributeText
 import org.codice.compliance.children
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_3_7
+import org.codice.compliance.Section.CORE_3_7
 import org.codice.compliance.utils.NodeDecorator
 import org.codice.compliance.utils.STATUS
 import org.codice.compliance.utils.STATUS_CODE
