+++ ctk/common/src/test/kotlin/org/codice/compilance/verification/core/responses/CoreAuthnRequestProtocolVerifierSpec.kt
@@ -15,7 +15,7 @@ import org.codice.compilance.ReportListener
 import org.codice.compliance.Common
 import org.codice.compliance.SAMLCore_3_3_2_2_1_a
 import org.codice.compliance.report.Report
-import org.codice.compliance.report.Report.Section.CORE_3_3
+import org.codice.compliance.Section.CORE_3_3
 import org.codice.compliance.utils.NodeDecorator
 import org.codice.compliance.utils.ddfAuthnContextList
 import org.codice.compliance.verification.core.responses.CoreAuthnRequestProtocolVerifier
