+++ library/src/main/kotlin/org/codice/compliance/SAMLSpecRefMessage.kt
@@ -8,31 +8,30 @@ http://www.gnu.org/licenses/lgpl.html
 
 package org.codice.compliance
 
-import org.codice.compliance.report.Report.Section
-import org.codice.compliance.report.Report.Section.BINDINGS_3_1
-import org.codice.compliance.report.Report.Section.BINDINGS_3_4
-import org.codice.compliance.report.Report.Section.BINDINGS_3_5
-import org.codice.compliance.report.Report.Section.CORE_1_3
-import org.codice.compliance.report.Report.Section.CORE_2_2
-import org.codice.compliance.report.Report.Section.CORE_2_3
-import org.codice.compliance.report.Report.Section.CORE_2_4
-import org.codice.compliance.report.Report.Section.CORE_2_5
-import org.codice.compliance.report.Report.Section.CORE_2_7
-import org.codice.compliance.report.Report.Section.CORE_3_2
-import org.codice.compliance.report.Report.Section.CORE_3_3
-import org.codice.compliance.report.Report.Section.CORE_3_4
-import org.codice.compliance.report.Report.Section.CORE_3_7
-import org.codice.compliance.report.Report.Section.CORE_4_1
-import org.codice.compliance.report.Report.Section.CORE_4_2
-import org.codice.compliance.report.Report.Section.CORE_5_4
-import org.codice.compliance.report.Report.Section.CORE_6_1
-import org.codice.compliance.report.Report.Section.CORE_8_2
-import org.codice.compliance.report.Report.Section.CORE_8_3
-import org.codice.compliance.report.Report.Section.GENERAL
-import org.codice.compliance.report.Report.Section.PROFILES_3_1
-import org.codice.compliance.report.Report.Section.PROFILES_4_1
-import org.codice.compliance.report.Report.Section.PROFILES_4_4
-import org.codice.compliance.report.Report.Section.SCHEMA
+import org.codice.compliance.Section.BINDINGS_3_1
+import org.codice.compliance.Section.BINDINGS_3_4
+import org.codice.compliance.Section.BINDINGS_3_5
+import org.codice.compliance.Section.CORE_1_3
+import org.codice.compliance.Section.CORE_2_2
+import org.codice.compliance.Section.CORE_2_3
+import org.codice.compliance.Section.CORE_2_4
+import org.codice.compliance.Section.CORE_2_5
+import org.codice.compliance.Section.CORE_2_7
+import org.codice.compliance.Section.CORE_3_2
+import org.codice.compliance.Section.CORE_3_3
+import org.codice.compliance.Section.CORE_3_4
+import org.codice.compliance.Section.CORE_3_7
+import org.codice.compliance.Section.CORE_4_1
+import org.codice.compliance.Section.CORE_4_2
+import org.codice.compliance.Section.CORE_5_4
+import org.codice.compliance.Section.CORE_6_1
+import org.codice.compliance.Section.CORE_8_2
+import org.codice.compliance.Section.CORE_8_3
+import org.codice.compliance.Section.GENERAL
+import org.codice.compliance.Section.PROFILES_3_1
+import org.codice.compliance.Section.PROFILES_4_1
+import org.codice.compliance.Section.PROFILES_4_4
+import org.codice.compliance.Section.SCHEMA
 import java.net.URI
 import java.util.ResourceBundle
